# app/openapi/swagger.yaml
openapi: 3.0.3
info:
  title: 'CCDI Data Federation: Participating Nodes API'
  description: |-
    This is the concrete OpenAPI specification for the CCDI
            Data Federation API—though this document contains all of the API calls
            that CCDI Federation nodes must implement, it does not outline the
            complete specification.

    * Visit the [documentation homepage](https://cbiit.github.io/ccdi-federation-api)
      to view the complete set of requirements to maintain and deploy a CCDI
      Federation node.
    * Additionally, you can view the Swagger specification in a more traditional
      theme by visiting
      [this link](https://editor.swagger.io/?url=https://cbiit.github.io/ccdi-federation-api/swagger.yml).
    * To access CCDI Data Federation Resource, please visit
      [this link](https://cbiit.github.io/ccdi-federation-api-aggregation/).
  contact:
    name: Childhood Cancer Data Initiative support email
    email: NCIChildhoodCancerDataInitiative@mail.nih.gov
  version: v1.2.0
servers:
- url: http://localhost:8000/api/v1
  description: DCC CCDI API server
paths:
...
  /sample:
    get:
      tags:
      - Sample
      summary: Gets the samples known by this server.
      description: |-
        Gets the samples known by this server.

        ### Pagination

        This endpoint is paginated. Users may override the default pagination
        parameters by providing one or more of the pagination-related query
        parameters below.

        ### Filtering

        All harmonized (top-level) and unharmonized (nested under the
        `metadata.unharmonized` key) metadata fields are filterable. To achieve
        this, you can provide the field name as a [`String`]. Filtering follows the
        following rules:

        * For single-value metadata field, the sample is included in the results if
        its value _exactly_ matches the query string. Matches are case-sensitive.
        * For multiple-value metadata fields, the sample is included in the results
        if any of its values for the field _exactly_ match the query string (a
        logical OR [`||`]). Matches are case-sensitive.
        * When the metadata field is `null` (in the case of singular or
        multiple-valued metadata fields) or empty, the sample is not included.
        * When multiple fields are provided as filters, a logical AND (`&&`) strings
        together the predicates. In other words, all filters must match for a
        sample to be returned. Note that this means that servers do not natively
        support logical OR (`||`) across multiple fields: that must be done by
        calling this endpoint with each of your desired queries and performing a
        set union of those samples out of band.

        ### Ordering

        This endpoint has default ordering requirements—those details are documented
        in the `responses::Samples` schema.
      operationId: sample_index
      parameters:
      - name: disease_phase
        in: query
        description: |-
          Matches any sample where the `disease_phase` field matches the string
          provided.
        required: false
        schema:
          type: string
      - name: anatomical_sites
        in: query
        description: |-
          Matches any sample where the `anatomical_sites` field matches the string
          provided.

          **Note:** a logical OR (`||`) is performed across the values when
          determining whether the subject should be included in the results.
        required: false
        schema:
          type: string
      - name: library_selection_method
        in: query
        description: |-
          Matches any sample where the `library_selection_method` field matches the string
          provided.
        required: false
        schema:
          type: string
      - name: library_strategy
        in: query
        description: |-
          Matches any sample where the `library_strategy` field matches the string
          provided.
        required: false
        schema:
          type: string
      - name: library_source_material
        in: query
        description: |-
          Matches any sample where the `library_source_material` field matches the string
          provided.
        required: false
        schema:
          type: string
      - name: preservation_method
        in: query
        description: |-
          Matches any sample where the `preservation_method` field matches the string
          provided.
        required: false
        schema:
          type: string
      - name: tumor_grade
        in: query
        description: |-
          Matches any sample where the `tumor_grade` field matches the string
          provided.
        required: false
        schema:
          type: string
      - name: specimen_molecular_analyte_type
        in: query
        description: |-
          Matches any sample where the `specimen_molecular_analyte_type` field matches the string
          provided.
        required: false
        schema:
          type: string
      - name: tissue_type
        in: query
        description: |-
          Matches any sample where the `tissue_type` field matches the string
          provided.
        required: false
        schema:
          type: string
      - name: tumor_classification
        in: query
        description: |-
          Matches any sample where the `tumor_classification` field matches the
          string provided.
        required: false
        schema:
          type: string
      - name: age_at_diagnosis
        in: query
        description: |-
          Matches any sample where the `age_at_diagnosis` field matches the string
          provided.
        required: false
        schema:
          type: string
      - name: age_at_collection
        in: query
        description: |-
          Matches any sample where the `age_at_collection` field matches the
          string provided.
        required: false
        schema:
          type: string
      - name: tumor_tissue_morphology
        in: query
        description: |-
          Matches any sample where the `tumor_tissue_morphology` field matches the
          string provided.
        required: false
        schema:
          type: string
      - name: depositions
        in: query
        description: |-
          Matches any sample where any member of the `depositions` fields match
          the string provided.

          **Note:** a logical OR (`||`) is performed across the values when
          determining whether the sample should be included in the results.
        required: false
        schema:
          type: string
      - name: diagnosis
        in: query
        description: |-
          Matches any sample where the `diagnosis` field matches the
          string provided.
        required: false
        schema:
          type: string
      - name: metadata.unharmonized.<field>
        in: query
        description: |-
          All unharmonized fields should be filterable in the same manner as harmonized fields:

          * Filtering on a singular field should include the `Sample` in the results if the query exactly matches the value of that field for the `Sample` (case-sensitive).
          * Filtering on field with multiple values should include the `Sample` in the results if the query exactly matches any of the values of the field for that `Sample` (case-sensitive).
          * Unlike harmonized fields, unharmonized fields must be prefixed with `metadata.unharmonized`.

          **Note:** this query parameter is intended to be symbolic of any unharmonized field. Because of limitations within Swagger UI, it will show up as a query parameter that can be optionally be submitted as part of a request within Swagger UI. Please keep in mind that the literal query parameter `?metadata.unharmonized.<field>=value` is not supported, so attempting to use it within Swagger UI will not work!
        required: false
        schema:
          type: string
      - name: page
        in: query
        description: |-
          The page to retrieve.

          This is a 1-based index of a page within a page set. The value of `page`
          **must** default to `1` when this parameter is not provided.
        required: false
        schema:
          type: integer
          minimum: 0
      - name: per_page
        in: query
        description: |-
          The number of results per page.

          Each server can select its own default value for `per_page` when this
          parameter is not provided. That said, the convention within the
          community is to use `100` as a default value if any value is equally
          reasonable.
        required: false
        schema:
          type: integer
          minimum: 0
      responses:
        '200':
          description: Successful operation.
          headers:
            link:
              schema:
                type: string
              description: "Links to URLs that may be of interest when paging through paginated responses. This header contains two or more links of interest. The format of the field is as follows: \n\n`Link: <URL>; rel=\"REL\"` \n### Relationships\n\nIn the format above, `URL` represents a valid URL for the link of interest and `REL` is one of four values: \n- `first` (_Required_). A link to the first page in the results (can be the same as `last` if there is only one page).\n- `last` (_Required_). A link to the first page in the results (can be the same as `first` if there is only one page).\n- `next` (_Optional_). A link to the next page (if it exists).\n- `prev` (_Optional_). A link to the previous page (if it exists).\n\n### Requirements\n\n- This header _must_ provide links for at least the `first` and `last` rels.\n - The `prev` and `next` links must exist only (a) when there are multiple pages in the result page set and (b) when the current page is not the first or last page, respectively.\n- This list of links is unordered.\n\n ### Notes\n\n- HTTP 1.1 and HTTP 2.0 dictate that response headers are case insensitive. Though not required, we recommend an all lowercase name of `link` for this response header."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responses.Samples'
        '404':
          description: |-
            Not found.
            Servers that cannot provide line-level data should use this response rather than Forbidden (403), as there is no level of authorization that would allow one to access the information included in the API.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responses.Errors'
              example:
                errors:
                - kind: UnshareableData
                  entity: Samples
                  reason: Our agreement with data providers prohibits us from sharing line-level data.
                  message: 'Unable to share data for samples: our agreement with data providers prohibits us from sharing line-level data.'
        '422':
          description: Invalid query or path parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responses.Errors'
              example:
                errors:
                - kind: InvalidParameters
                  parameters:
                  - page
                  - per_page
                  reason: Unable to calculate offset.
                  message: 'Invalid value for parameters ''page'' and ''per_page'': unable to calculate offset.'
  /sample/{organization}/{namespace}/{name}:
    get:
      tags:
      - Sample
      summary: Gets the sample matching the provided name (if the sample exists).
      description: Gets the sample matching the provided name (if the sample exists).
      operationId: sample_show
      parameters:
      - name: organization
        in: path
        description: The organization identifier of the namespace to which the sample belongs.
        required: true
        schema:
          type: string
      - name: namespace
        in: path
        description: The name of the namespace to which the sample belongs.
        required: true
        schema:
          type: string
      - name: name
        in: path
        description: The name portion of the sample identifier.
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responses.Sample'
        '404':
          description: |-
            Not found.
            Servers that cannot provide line-level data should use this response rather than Forbidden (403), as there is no level of authorization that would allow one to access the information included in the API.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responses.Errors'
              example:
                errors:
                - kind: NotFound
                  entity: Sample with namespace 'foo' and name 'bar'
                  message: Sample with namespace 'foo' and name 'bar' not found.
  /sample/by/{field}/count:
    get:
      tags:
      - Sample
      summary: Groups the samples by the specified metadata field and returns counts.
      description: Groups the samples by the specified metadata field and returns counts.
      operationId: samples_by_count
      parameters:
      - name: field
        in: path
        description: The field to group by and count with.
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responses.by.count.sample.Results'
        '422':
          description: Unsupported field.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responses.Errors'
              example:
                errors:
                - kind: UnsupportedField
                  field: handedness
                  reason: This field is not present for samples.
                  message: 'Field ''handedness'' is not supported: this field is not present for samples.'
...
components:
  schemas:
...
    models.Sample:
      type: object
      description: |-
        A sample.

        **Note:** the `subject` identifier **must** match a
        [`Subject`](super::Subject) that both (a) is listed in the
        [`Subject`](super::Subject) index endpoint and (b) is able to be shown with
        the [`Subject`](super::Subject) show endpoint.
      required:
      - id
      - subject
      properties:
        id:
          $ref: '#/components/schemas/models.sample.Identifier'
        subject:
          $ref: '#/components/schemas/models.subject.Identifier'
        gateways:
          type: array
          items:
            $ref: '#/components/schemas/models.gateway.AnonymousOrReference'
          description: |-
            One or more [gateways](AnonymousOrReference) through which this sample
            can be accessed.

            Gateways can be either [anonymous](AnonymousOrReference::Anonymous)
            ([gateways](crate::Gateway) with no name) or a
            [refererence](AnonymousOrReference::Reference) to a [named
            gateway](gateway::Named) ([gateways](crate::Gateway) with a name).

            **Anonymous** gateways are intended to be embedded directly within a
            returned sample in the `/sample` response object. They have no name and
            are only referred to by the sample within which they are embedded.

            **Named** gateways, on the other hand, are included in the `gateways`
            key in the `/sample` response object and referred to by name within a
            returned sample in the `/sample` response object. They are intended to
            be used when more than one sample references the same gateway. This
            mechanism is available to ensure that the gateway object does not need
            to be duplicated multiple times in the response in these cases.

            This field can contain multiple gateways to support scenarios where a
            sample is available through more than one mechanism. We expect that only
            one gateway will be returned in most responses (if at all).
        metadata:
          allOf:
          - $ref: '#/components/schemas/models.sample.Metadata'
          nullable: true
...
    models.gateway.AnonymousOrReference:
      oneOf:
      - type: object
        description: An anonymous gateway.
        required:
        - gateway
        - kind
        properties:
          gateway:
            $ref: '#/components/schemas/models.Gateway'
          kind:
            type: string
            enum:
            - Anonymous
      - type: object
        description: A reference to a named gateway.
        required:
        - gateway
        - kind
        properties:
          gateway:
            type: string
            description: The reference to a [`Named`] gateway.
          kind:
            type: string
            enum:
            - Reference
      description: An anonymous [`Gateway`] or a reference to a named [`Gateway`].
      discriminator:
        propertyName: kind
...
    models.gateway.Named:
      allOf:
      - $ref: '#/components/schemas/models.Gateway'
      - type: object
        required:
        - name
        properties:
          name:
            type: string
            description: The name.
      description: |-
        A named gateway.

        A named gateway is simply a [`Gateway`] with a name. Named gateways exist so
        that multiple files in a response can refer to the same gateway (by name)
        without duplicating the information for that gateway multiple times.

        **Note:** a _named_ gateway can only be included in a `gateways` response
        object—they cannot be embedded directly within a [`File`](crate::File) in
        the response.
...
    models.namespace.Identifier:
      type: object
      description: An identifier for a namespace.
      required:
      - organization
      - name
      properties:
        organization:
          $ref: '#/components/schemas/models.organization.Identifier'
        name:
          $ref: '#/components/schemas/models.namespace.identifier.Name'
...
    models.sample.Identifier:
      type: object
      description: |-
        An identifier for a [`Sample`](crate::Sample).

        [`Identifiers`](Identifier) serve two main purposes:

        1. They represent the primary identifier for a [`Sample`](crate::Sample).
        2. They extended when referenced as [linked identifiers](linked::Identifier).
      required:
      - namespace
      - name
      properties:
        namespace:
          $ref: '#/components/schemas/models.namespace.Identifier'
        name:
          type: string
          description: |-
            **`caDSR CDE 15100774 v1.00`**

            This metadata element is defined by the caDSR as "A unique string of characters
            used to identify a specimen.".

            Link:
            <https://cadsr.cancer.gov/onedata/dmdirect/NIH/NCI/CO/CDEDD?filter=CDEDD.ITEM_ID=15100774%20and%20ver_nr=1>
          example: SampleName001
    models.sample.Metadata:
      allOf:
      - $ref: '#/components/schemas/models.metadata.common.Metadata'
      - type: object
        required:
        - age_at_diagnosis
        - anatomical_sites
        - diagnosis
        - disease_phase
        - library_selection_method
        - tissue_type
        - tumor_classification
        - tumor_tissue_morphology
        - age_at_collection
        - library_strategy
        - library_source_material
        - preservation_method
        - tumor_grade
        - specimen_molecular_analyte_type
        - identifiers
        properties:
          age_at_diagnosis:
            allOf:
            - $ref: '#/components/schemas/field.unowned.sample.AgeAtDiagnosis'
            nullable: true
          anatomical_sites:
            allOf:
            - $ref: '#/components/schemas/field.unowned.sample.AnatomicalSite'
            nullable: true
          diagnosis:
            allOf:
            - $ref: '#/components/schemas/field.unowned.sample.Diagnosis'
            nullable: true
          disease_phase:
            allOf:
            - $ref: '#/components/schemas/field.unowned.sample.DiseasePhase'
            nullable: true
          library_selection_method:
            allOf:
            - $ref: '#/components/schemas/field.unowned.sample.LibrarySelectionMethod'
            nullable: true
          tissue_type:
            allOf:
            - $ref: '#/components/schemas/field.unowned.sample.TissueType'
            nullable: true
          tumor_classification:
            allOf:
            - $ref: '#/components/schemas/field.unowned.sample.TumorClassification'
            nullable: true
          tumor_tissue_morphology:
            allOf:
            - $ref: '#/components/schemas/field.unowned.sample.TumorTissueMorphology'
            nullable: true
          age_at_collection:
            allOf:
            - $ref: '#/components/schemas/field.unowned.sample.AgeAtCollection'
            nullable: true
          library_strategy:
            allOf:
            - $ref: '#/components/schemas/field.unowned.sample.LibraryStrategy'
            nullable: true
          library_source_material:
            allOf:
            - $ref: '#/components/schemas/field.unowned.sample.LibrarySourceMaterial'
            nullable: true
          preservation_method:
            allOf:
            - $ref: '#/components/schemas/field.unowned.sample.PreservationMethod'
            nullable: true
          tumor_grade:
            allOf:
            - $ref: '#/components/schemas/field.unowned.sample.TumorGrade'
            nullable: true
          specimen_molecular_analyte_type:
            allOf:
            - $ref: '#/components/schemas/field.unowned.sample.SpecimenMolecularAnalyteType'
            nullable: true
          identifiers:
            type: array
            items:
              $ref: '#/components/schemas/field.unowned.sample.Identifier'
            description: |-
              The alternate identifiers for the sample.

              Note that this list of identifiers *must* include the main identifier
              for the [`Sample`].
            nullable: true
          unharmonized:
            $ref: '#/components/schemas/fields.Unharmonized'
      description: Metadata associated with a sample.
...
    models.subject.Identifier:
      type: object
      description: |-
        An identifier for a [`Subject`](crate::Subject).

        [`Identifiers`](Identifier) serve two main purposes:

        1. They represent the primary identifier for a [`Subject`](crate::Subject).
        2. They extended when referenced as [linked identifiers](linked::Identifier).
      required:
      - namespace
      - name
      properties:
        namespace:
          $ref: '#/components/schemas/models.namespace.Identifier'
        name:
          $ref: '#/components/schemas/cde.v1.subject.Name'
...
    responses.Errors:
      type: object
      description: A wrapper around one or more [errors](Kind).
      required:
      - errors
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/responses.error.Kind'
          description: The errors within this response.
...
    responses.Sample:
      allOf:
      - $ref: '#/components/schemas/models.Sample'
      - type: object
      description: A response representing a single [`Sample`](models::Sample).
    responses.Samples:
      type: object
      description: |-
        A response representing multiple samples known about by the server.

        When no sort order is provided, samples **must** be ordered by the primary
        identifier. This means that, when comparing two identifiers:

        1. The namespace organization field should be sorted alphabetically. If all
        values for the namespace organization are equal, continue on to the next
        sorting criteria.
        2. The namespace name field should be sorted alphabetically. If all values
        for the namespace names are equal, continue on to the next sorting
        criteria.
        3. The entity name should be sorted alphabetically.

        Since the `namespace` and `name` identifiers should always uniquely apply to
        a single entity, this should always resolve to an ordering.

        If there is a provided sort order, use that instead.
      required:
      - summary
      - data
      properties:
        summary:
          $ref: '#/components/schemas/responses.entity.Summary'
        data:
          type: array
          items:
            $ref: '#/components/schemas/models.Sample'
          description: The samples.
        gateways:
          type: array
          items:
            $ref: '#/components/schemas/models.gateway.Named'
...
    responses.by.count.ValueCount:
      type: object
      description: A value along with the number of counted entities for that value.
      required:
      - value
      - count
      properties:
        value:
          description: The value.
        count:
          type: integer
          description: The number of times the value was counted.
          minimum: 0
...
    responses.by.count.sample.Results:
      type: object
      description: |-
        A set of results from grouping [`Samples`](ccdi_models::Sample) by a specified
        metadata field and then summing the counts for each value (along with computing a
        total count).
      required:
      - total
      - missing
      - values
      properties:
        total:
          type: integer
          description: The total number of counts in this result set.
          minimum: 0
        missing:
          type: integer
          description: |-
            The total number of entries that are missing values. In this context,
            "missing" means either (a) the individual metadata key is missing or (b)
            the entire metadata object is missing.
          minimum: 0
        values:
          type: array
          items:
            $ref: '#/components/schemas/responses.by.count.ValueCount'
          description: The counts per value observed for the result set.
...
    responses.entity.Counts:
      type: object
      description: Counts that summarize the contents of a paged entity response.
      required:
      - current
      - all
      properties:
        current:
          type: integer
          description: The number of entities within the currently selected page in the result set.
          minimum: 0
        all:
          type: integer
          description: The number of entities across all pages in the result set.
          minimum: 0
    responses.entity.Summary:
      type: object
      description: A summary of a paged entity response.
      required:
      - counts
      properties:
        counts:
          $ref: '#/components/schemas/responses.entity.Counts'
    responses.error.Kind:
      allOf:
      - oneOf:
        - type: object
          description: |-
            Attempted to access an invalid route.

            Also includes all routes for which the path exists, but the HTTP method
            is not supported for that path.
          required:
          - method
          - route
          - kind
          properties:
            method:
              type: string
              description: The HTTP method that was used in the request.
            route:
              type: string
              description: The route that was requested.
            kind:
              type: string
              enum:
              - InvalidRoute
          example:
            kind: InvalidRoute
            method: GET
            route: /foobar
        - type: object
          description: One or more invalid query or path parameters were provided.
          required:
          - reason
          - kind
          properties:
            parameters:
              type: array
              items:
                type: string
              description: |-
                If known, the parameters that are invalid. If not known or the list
                of invalid fields is not desirable to share, pass `null` to this
                field for a more general error message.
              nullable: true
            reason:
              type: string
              description: A plain-text reason describing why the parameters are invalid.
            kind:
              type: string
              enum:
              - InvalidParameters
          example:
            kind: InvalidParameters
            parameters:
            - id
            reason: The parameter was a non-integer value.
        - type: object
          description: An entity was not found.
          required:
          - entity
          - kind
          properties:
            entity:
              type: string
              description: The entity (or entities) that are not found.
            kind:
              type: string
              enum:
              - NotFound
          example:
            kind: NotFound
            entity: Samples
        - type: object
          description: Line-level data cannot be shared for the specified entity.
          required:
          - entity
          - reason
          - kind
          properties:
            entity:
              type: string
              description: The entity (or entities) where data cannot be shared.
            reason:
              type: string
              description: The reason that the line-level data cannot be shared.
            kind:
              type: string
              enum:
              - UnshareableData
          example:
            kind: UnshareableData
            entity: Sample
            reason: Our agreement with data providers prohibits us from sharing line-level data.
        - type: object
          description: A field name was not supported for the attempted operation.
          required:
          - field
          - reason
          - kind
          properties:
            field:
              type: string
              description: The field that is not supported.
            reason:
              type: string
              description: The reason that the field is not supported.
            kind:
              type: string
              enum:
              - UnsupportedField
          example:
            kind: UnsupportedField
            field: field
            reason: The field was not found in the metadata object.
        description: An inner type of a [Kind](super::Kind).
        discriminator:
          propertyName: kind
      - type: object
        required:
        - message
        properties:
          message:
            type: string
            description: |-
              A plain-text description of the error.

              This field is intended to be shown within a user interface or similar if
              needed. Please use this field if you intend to pass the error along to a
              user.
      description: A response indicating an error from the API.
...